+++ src/runtime/__tests__/arena-scalar-write-through.test.ts
@@ -50,7 +50,7 @@ function createStateForProgram(program: CompiledProgramIR): RuntimeState {
   const schedule = program.schedule as ScheduleIR;
   const sizes = computeStorageSizes(program.slotMeta);
   return createRuntimeState(
-    sizes.f64,
+    sizes.f32,
     schedule.stateSlotCount ?? 0,
     schedule.eventSlotCount ?? 0,
     schedule.eventCount ?? 0,
